@@ -11,9 +11,10 @@
     <description>The core plugin for VRN Network.</description>
 
     <properties>
-        <foundation.version>5.8.3</foundation.version>
-        <java.version>1.8</java.version>
+        <java.version>16</java.version>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
+        <mc.version>1.17</mc.version>
+        <vrnlib>1.2a</vrnlib>
     </properties>
 
     <repositories>
