@@ -31,21 +31,21 @@ public class WarpsGUI extends PageableGUI<Warp> {
     }
 
     @Override
-    protected ItemStack convertToItem(final Warp w) {
-        final Block b = VRNUtil.getStandingBlock(w.getLocation());
-        return new ItemBuilder(b != null ? b.getType() : Material.BARRIER).setName("&7" + w.getName());
+    protected ItemStack convertToItem(Warp w) {
+        Block b = VRNUtil.getStandingBlock(w.location());
+        return new ItemBuilder(b != null ? b.getType() : Material.BARRIER).setName("&7" + w.name());
     }
 
     @Override
-    protected void onClickItem(final Warp w, final InventoryClickEvent e) {
-        final Player player = getViewer();
+    protected void onClickItem(Warp w, InventoryClickEvent e) {
+        Player player = getViewer();
         if (e.getClick().isLeftClick()) {
             player.closeInventory();
             say(player, Messages.msg("teleporting"));
-            player.teleport(w.getLocation());
+            player.teleport(w.location());
         }
         if (e.getClick().isRightClick()) {
-            final String perm = "vrn.delwarp." + (!w.getOwner().equals(player.getUniqueId()) ? "others" : "self");
+            String perm = "vrn.delwarp." + (!w.owner().equals(player.getUniqueId()) ? "others" : "self");
             if (!player.hasPermission(perm)) {
                 getViewer().closeInventory();
                 say(getViewer(), VRNUtil.noperm);
@@ -63,14 +63,14 @@ public class WarpsGUI extends PageableGUI<Warp> {
 
     private static class DeleteConfirm {
 
-        private DeleteConfirm(final Player player, final Warp w) {
-            final InventoryGUI gui = new InventoryGUI(9, "&c&lConfirm delete?");
+        private DeleteConfirm(Player player, Warp w) {
+            InventoryGUI gui = new InventoryGUI(9, "&c&lConfirm delete?");
             gui.addButton(ItemButton.create(new ItemBuilder(Material.RED_WOOL).setName("&cCancel (Back to homes list)"), e -> new WarpsGUI().open(player)), 2);
             gui.getInventory().setItem(4, new ItemBuilder(Material.MAP).setName("&6Are you sure?").setLore("", "&eAre you sure you want to", "&edelete this home?"));
             gui.addButton(ItemButton.create(new ItemBuilder(Material.LIME_WOOL).setName("&aConfirm"), e -> {
                 VRNcore.getInstance().getWarpManager().deleteWarp(w);
                 player.closeInventory();
-                say(player, "&7Warp &a" + w.getName() + "&7 successfully deleted.");
+                say(player, "&7Warp &a" + w.name() + "&7 successfully deleted.");
             }), 6);
             gui.open(player);
         }
