@@ -1,7 +1,7 @@
 package net.skeagle.vrncore.event;
 
 import net.skeagle.vrncore.VRNcore;
-import net.skeagle.vrncore.trail.VRNParticle;
+import net.skeagle.vrncore.trail.Trails;
 import net.skeagle.vrncore.utils.VRNPlayer;
 import org.bukkit.Color;
 import org.bukkit.Particle;
@@ -15,13 +15,13 @@ import org.bukkit.scheduler.BukkitRunnable;
 public class ArrowListener implements Listener {
 
     @EventHandler
-    public void onArrowShot(final ProjectileLaunchEvent e) {
-        final VRNPlayer vrnPlayer;
+    public void onArrowShot(ProjectileLaunchEvent e) {
+        VRNPlayer vrnPlayer;
         if (e.getEntity().getShooter() instanceof Player && e.getEntity() instanceof Arrow) {
             vrnPlayer = new VRNPlayer((Player) e.getEntity().getShooter());
-            final Particle particle = vrnPlayer.getArrowTrail();
+            Particle particle = vrnPlayer.getArrowTrail();
             if (particle != null) {
-                final String perm = VRNParticle.getNameFromParticle(particle);
+                String perm = Trails.getNameFromParticle(particle);
                 if (perm != null) {
                     if (((Player) e.getEntity().getShooter()).hasPermission("vrn.arrowtrails." + perm)) {
                         new BukkitRunnable() {
