@@ -17,59 +17,51 @@ public final class VRNUtil {
 
     public static String noperm = color(Messages.getLoaded(VRNLib.getInstance()).get("noPermission"));
 
-    public static void say(final CommandSender cs, final String... message) {
+    public static void say(CommandSender cs, String... message) {
         if (cs == null) return;
-        for (final String msg : message)
+        for (String msg : message)
             cs.sendMessage(color(Messages.msg("prefix") + " " + msg));
     }
 
-    public static void say(final VRNPlayer p, final String... message) {
-        for (final String msg : message)
+    public static void say(VRNPlayer p, String... message) {
+        for (String msg : message)
             p.getPlayer().sendMessage(color(Messages.msg("prefix") + " " + msg));
     }
 
-    public static void sayNoPrefix(final CommandSender cs, final String... message) {
+    public static void sayNoPrefix(CommandSender cs, String... message) {
         if (cs == null) return;
-        for (final String msg : message)
+        for (String msg : message)
             cs.sendMessage(color(msg));
     }
 
-    public static void sayNoPrefix(final VRNPlayer p, final String... message) {
-        for (final String msg : message)
+    public static void sayNoPrefix(VRNPlayer p, String... message) {
+        for (String msg : message)
             p.getPlayer().sendMessage(color(msg));
     }
 
-    public static boolean hasPerm(final Player p, final String perm) {
-        return p.hasPermission(perm);
-    }
-
-    public static boolean hasPerm(final VRNPlayer p, final String perm) {
+    public static boolean hasPerm(VRNPlayer p, String perm) {
         return p.getPlayer().hasPermission(perm);
     }
 
-    public static <T extends Exception> void sneakyThrow(final Throwable t) throws T {
-        throw (T) t;
-    }
-
-    public static void log(final String... messages) {
+    public static void log(String... messages) {
         log(Level.INFO, messages);
     }
 
-    public static void log(final Level level, final String... messages) {
+    public static void log(Level level, String... messages) {
         if (messages == null) return;
-        for (final String s : messages)
+        for (String s : messages)
             Bukkit.getLogger().log(level, color(s));
     }
 
     public static String color(String s) {
         for (int i = 0; i < s.length(); ++i) {
             if (s.length() - i > 8) {
-                final String temp = s.substring(i, i + 8);
+                String temp = s.substring(i, i + 8);
                 if (temp.startsWith("&#")) {
-                    final char[] chars = temp.replaceFirst("&#", "").toCharArray();
-                    final StringBuilder rgbColor = new StringBuilder();
+                    char[] chars = temp.replaceFirst("&#", "").toCharArray();
+                    StringBuilder rgbColor = new StringBuilder();
                     rgbColor.append("&x");
-                    for (final char c : chars)
+                    for (char c : chars)
                         rgbColor.append("&").append(c);
                     s = s.replaceAll(temp, rgbColor.toString());
                 }
@@ -78,30 +70,30 @@ public final class VRNUtil {
         return ChatColor.translateAlternateColorCodes('&', s);
     }
 
-    public static String[] color(final String... i) {
-        for (final String uncolored : i)
+    public static String[] color(String... i) {
+        for (String uncolored : i)
             color(uncolored);
         return i;
     }
 
-    public static void sayActionBar(final Player p, final String msg) {
+    public static void sayActionBar(Player p, String msg) {
         p.spigot().sendMessage(ChatMessageType.ACTION_BAR,
                 TextComponent.fromLegacyText(color(msg)));
     }
 
-    public static Block getStandingBlock(final Location loc) {
-        final Location l = loc.clone();
-        final Block standing = l.add(0, -1, 0).getBlock();
-        final double x = l.getX() - (double) l.getBlockX();
-        final double z = l.getZ() - (double) l.getBlockZ();
-        final Block b1 = standing.getLocation().clone().add(1, 0, 0).getBlock();
-        final Block b2 = standing.getLocation().clone().add(1, 0, 1).getBlock();
-        final Block b3 = standing.getLocation().clone().add(0, 0, 1).getBlock();
-        final Block b4 = standing.getLocation().clone().add(-1, 0, 1).getBlock();
-        final Block b5 = standing.getLocation().clone().add(-1, 0, 0).getBlock();
-        final Block b6 = standing.getLocation().clone().add(-1, 0, -1).getBlock();
-        final Block b7 = standing.getLocation().clone().add(0, 0, -1).getBlock();
-        final Block b8 = standing.getLocation().clone().add(1, 0, -1).getBlock();
+    public static Block getStandingBlock(Location loc) {
+        Location l = loc.clone();
+        Block standing = l.add(0, -1, 0).getBlock();
+        double x = l.getX() - (double) l.getBlockX();
+        double z = l.getZ() - (double) l.getBlockZ();
+        Block b1 = standing.getLocation().clone().add(1, 0, 0).getBlock();
+        Block b2 = standing.getLocation().clone().add(1, 0, 1).getBlock();
+        Block b3 = standing.getLocation().clone().add(0, 0, 1).getBlock();
+        Block b4 = standing.getLocation().clone().add(-1, 0, 1).getBlock();
+        Block b5 = standing.getLocation().clone().add(-1, 0, 0).getBlock();
+        Block b6 = standing.getLocation().clone().add(-1, 0, -1).getBlock();
+        Block b7 = standing.getLocation().clone().add(0, 0, -1).getBlock();
+        Block b8 = standing.getLocation().clone().add(1, 0, -1).getBlock();
         //check direct block
         if (blockCheck(standing)) {
             return standing;
@@ -136,20 +128,20 @@ public final class VRNUtil {
         return null;
     }
 
-    private static boolean blockCheck(final Block b) {
+    private static boolean blockCheck(Block b) {
         return (!b.getType().isAir() && b.getType().isSolid());
     }
 
     public static class LocationSerialization {
-        public static String serialize(final Location loc) {
+        public static String serialize(Location loc) {
             if (loc == null || loc.getWorld() == null)
                 return null;
             return loc.getWorld().getName() + " " + loc.getX() + " " + loc.getY() + " " + loc.getZ() + " " + loc.getYaw() + " " + loc.getPitch();
         }
 
-        public static Location deserialize(final String s) {
+        public static Location deserialize(String s) {
             if (s == null) return null;
-            final String[] split = s.split(" ");
+            String[] split = s.split(" ");
             if (split.length != 6)
                 return null;
             return new Location(Bukkit.getWorld(split[0]), Double.parseDouble(split[1]), Double.parseDouble(split[2]), Double.parseDouble(split[3]),
